WIP for async diffs
