+++ indigo_lib/__init__.py
@@ -1,0 +1,3 @@
+# indigo_lib has routines that are entirely independent of Django, and can be loaded without initialising Django.
+#
+# NOTE: do NOT reference any Django models or code in this module, or import any modules that reference Django models
